Move SFINAE helpers along the definitions of device/host unique_ptrs
